streamlit-app/requirements.txt
streamlit==1.47.1
pillow==11.3.0
torch==2.7.1
torchvision==0.22.1